requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.38.0
    # via matplotlib
fuzzywuzzy==0.18.0
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
levenshtein==0.20.8
    # via python-levenshtein
matplotlib==3.6.2
    # via -r requirements.in
numpy==1.23.5
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==21.3
    # via matplotlib
pillow==9.3.0
    # via matplotlib
pydantic==1.10.2
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
python-levenshtein==0.20.8
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
rapidfuzz==2.13.2
    # via levenshtein
six==1.16.0
    # via python-dateutil
typing-extensions==4.4.0
    # via pydantic
